requirements.txt
Adafruit-DHT==1.4.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cryptography==1.7.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
gspread==3.1.0
httplib2==0.12.3
idna==2.8
keyring==10.1
keyrings.alt==1.3
oauth2client==4.1.3
oauthlib==3.0.1
paho-mqtt==1.4.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycrypto==2.6.1
pygobject==3.22.0
python-apt==1.1.0b5
pyxdg==0.25
requests==2.21.0
requests-oauthlib==1.2.0
RPi.GPIO==0.6.5
rsa==4.0
SecretStorage==2.3.1
six==1.12.0
slackclient==1.3.1
ssh-import-id==5.6
uritemplate==3.0.0
urllib3==1.24.3
websocket-client==0.54.0
